refactor(suite-desktop): move modules loader into "modules" directory
