+++ packages/suite-desktop/src-electron/modules/tor.ts
@@ -5,7 +5,7 @@ import { session } from 'electron';
 import TorProcess, { DEFAULT_ADDRESS } from '../libs/processes/TorProcess';
 import { onionDomain } from '../config';
 import { app, ipcMain } from '../typed-electron';
-import { Module } from '../libs/modules';
+import { Module } from './index';
 
 const init: Module = async ({ mainWindow, store, interceptor }) => {
     const { logger } = global;
